refactor : viewerpage hook 분리
